Add Hand-coded GPU draw string function
Can be used to draw a string on-screen using a
hard-coded font. Use this to print the same
GPU and DSP debugging information we printf to
the skunk or GameDrive debug streams on the screen
for testing purposes. Modify the Makefile to
always build sprintf() support in even when not
using the skunk or GameDrive libraries, since it's
useful for on-screen text now too.