Add animation system
So far this just moves sprites at a fixed speed
per frame. Use it to move the "player" sprite
a certain number of pixels on each button press
at a fixed speed.

This functionality had to be hand-coded due to
bugs in the C compiler. The hand-code is derived
from the mostly-working output of the comiler,
with some conditional logic fixed and some bugs
from the original C code fixed.