Use .regequ to accomodate different C compilers
Different builds of the Atari/Brainstorm GCC JRISC
compiler use different registers for the stack
pointer. Make it easier to switch between the two
by adding some .regequ aliases equivalent to those
output in the GCC boilerplate and using them in
the stack management portions of the hand-coded
assembly functions.