Optimize pick_numbers() for size.
Take advantage of the fact that this function
only calls get_rand_entry(), and that it is the
only caller of that function to safely make use
of local rather than global registers and hence
eliminate all of the code used to save & restore
the global registers it previously used.

This reduces the total GPU code size by 60 bytes!