Split gpugame.c in two
The minimal first step to split the C GPU code
into two separate C files. The goal is that the
first file will be a core set of helper functions
used by various overlays, each defined in their
own C file. For now, there is no functional change
and the generated code should be essentially
identical, but it is now generated and loaded as
two separate modules, similar to how the GPU ASM
and C code were modularized before.