Load overlays on the GPU
Modify the GPU common start function to handle
loading overlays beyond the initial common code.

Copy the blitGpuOverlay function and its helper
to gpucommon.c, adjusting it to reference the
blitter registers as volatile pointers to avoid
the optimizer reordering register writes, and to
use the BLITTER_WAIT() helper.

Have each function modify the global variable
nextOverlay to indicate whether it has a follow-
on overlay. If it does not, stop the GPU after
it runs in the GPU common startup function.