Use max of each GPU code overlay as ASM start addr
Right now the level init code is larger than the
in-level game logic. but the whole point of this
split was to give the game logic room to grow.
Rather than bake in assumptions about which GPU C
code overlay is larger, calculate the max as part
of building the GPU ASM hand code object using
RMAC assembler conditional logic.